pyproject.toml
[tool.poetry]
name = "cars"
version = "0.1.0"
description = ""
authors = ["silverfatt <silvestr.fat@mail.ru>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.115.6"
uvicorn = "^0.34.0"
pydantic-settings = "^2.7.0"
loguru = "^0.7.3"
asyncpg = "^0.30.0"
python-jose = "^3.3.0"
passlib = "^1.7.4"
python-multipart = "^0.0.20"
bcrypt = "^4.2.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
